backend/requirements.txt
flask==2.0.1
face-recognition==1.3.0
numpy==1.21.2
opencv-python==4.5.3.56
pillow==8.3.1
python-dotenv==0.19.0